Removing non-unittest test_files, and moving test_grid_gradient to test_grid
